backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
aiosqlite==0.20.0
sse-starlette==2.2.1
mcp>=1.0.0
e2b>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
fpdf2>=2.7.0
